fix: Force reload of pages stored in bfcache if they are revisited
After setting the Cache-Control header to prevent caching/storing of
webpages, I noticed that some pages were still in fact being cached
during testing. This is what I discovered:

- If a user successfully signs up, they are redirected to the login page
(which is what I want)
- On the login page, if the user hits the back button, they will go back
to the signup page, and the credentials they entered will still be
displayed in the input fields. I do not want this behavior to happen due
to security concerns, which is why I set Cache-Control to no-store in
the first place.

After doing much research I discovered that these pages were being
stored in bfcache, which appears to be sometimes unaffected by the
Cache-Control header. To fix this issue I added event listeners for
'pageshow' and 'pagehide' on pages that contain a form, and when
those events are fired they will either forcefully reload the page
or replace the current page's state in the tab's history.